refactor: 랭킹 정보 생성에 불필요한 부분 제거

+++ src/main/kotlin/com/sat/user/query/PointQuery.kt
@@ -2,6 +2,6 @@ package com.sat.user.query
 
 data class PointQuery(
     val memberId: Long,
-    val point: Double,
+    val point: Int,
     var memberName: String? = null,
 )
